Optimise projects route with Project component
